requirements copy.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
arrow==1.3.0
attrs==24.3.0
binaryornot==0.4.4
cachetools==5.5.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
defusedxml==0.7.1
deprecation==2.1.0
fastapi==0.115.6
flet==0.25.2
flet-cli==0.25.2
flet-desktop==0.25.2
flet-web==0.25.2
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
gotrue==2.11.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
postgrest==0.18.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
realtime==2.0.6
repath==0.9.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==4.0.2
watchfiles==1.0.3
websockets==13.1
yarl==1.18.3
youtube-transcript-api==0.6.3
pyinstaller